Began moving ImageSubmit to seperate files
